Added exception handling and tested all endpoints in controller

+++ services/user-service/src/main/java/com/tkforgeworks/cookconnect/userservice/service/CCUserService.java
@@ -3,6 +3,7 @@ package com.tkforgeworks.cookconnect.userservice.service;
 import com.tkforgeworks.cookconnect.userservice.model.CCUser;
 import com.tkforgeworks.cookconnect.userservice.model.dto.CCUserDto;
 import com.tkforgeworks.cookconnect.userservice.model.dto.NoProfileCCUserDTO;
+import com.tkforgeworks.cookconnect.userservice.model.dto.PasswordDto;
 import com.tkforgeworks.cookconnect.userservice.model.dto.UpdateCCUserDTO;
 import com.tkforgeworks.cookconnect.userservice.model.mapper.UserServiceMapper;
 import com.tkforgeworks.cookconnect.userservice.repository.CCUserRepository;
@@ -63,7 +64,7 @@ public class CCUserService {
         if(!Objects.equals(updateCCUserDTO.id(), ccUserId)){
             //TODO: BadRequestError
             log.error("User with id {} does not match id {}", updateCCUserDTO.id(), ccUserId);
-            return null;
+            throw new RuntimeException(String.format("User with id %s does not match id %s", updateCCUserDTO.id(), ccUserId));
         }
         CCUser userToUpdate = userRepository.findById(updateCCUserDTO.id())
                 .orElseThrow(() -> new RuntimeException("User not found"));
@@ -75,7 +76,25 @@ public class CCUserService {
 
     public String deleteUserById(Long ccUserId) {
         //TODO: add ifExists check prior to delete for better response message
+        if(!userRepository.existsById(ccUserId)){
+            log.error("User with id {} does not exists", ccUserId);
+            throw new RuntimeException(String.format("User with id %s not found", ccUserId));
+        }
         userRepository.deleteById(ccUserId);
         return "User with id " + ccUserId + " deleted";
+    }
+
+    public void addNewPassword(Long ccUserId, PasswordDto password) {
+        CCUser foundUser = userRepository.findById(ccUserId).orElseThrow(()-> new RuntimeException("user not found"));
+        if(!Objects.equals(foundUser.getUsername(), password.username())) {
+            log.error("User with id {} does not match username passed: {}", ccUserId, password.username());
+            throw new RuntimeException(String.format("User with id %s does not match username passed: %s", ccUserId, password.username()));
+        }
+        if(Objects.nonNull(foundUser.getPassword()) && !foundUser.getPassword().isEmpty()) {
+            log.error("User {} already has password set", foundUser.getUsername());
+            throw new RuntimeException(String.format("User with id %s already has password set", foundUser.getUsername()));
+        }
+        foundUser.setPassword(password.password());
+        userRepository.save(foundUser);
     }
 }
